refactor(signer): move userPreferences helper to _shared/, drop cron duplicate
Each Supabase edge function has its own Deno module graph and can't
import from a sibling function's lib/. Lane B created
farcaster-signer/lib/userPreferences.ts and the cron then maintained
an inline mirror (getUserFarcasterProviderForCron) with the same
query + validation. Codex flagged the drift risk in the second
review pass.

Moves the helper to supabase/functions/_shared/userPreferences.ts
(matches the existing pattern used by mentions.ts and redact.ts).
HubProvider type lives there too — single source of truth — and
farcaster-signer/lib/hubs.ts re-exports it so local imports
(`import { HubProvider } from './hubs.ts'`) keep working unchanged.

Uses a minimal structural SupabaseLikeClient interface inside the
shared file so we don't have to pick a Deno import specifier
(`npm:`, bare, or `https://esm.sh/`) — each caller passes their
own typed client and structural typing matches both.

Net: ~80 lines deleted, ~50 lines added. No behavior change.
